move user house fetch to house provider
